requirements.txt
praw>=7.8.1
anthropic>=0.40.0
numpy>=2.0.0
pandas>=2.3.0
python-dotenv>=1.0.0
python-docx>=1.1.0
matplotlib>=3.9.0
seaborn>=0.13.0
jupyterlab>=4.0.0
wordcloud>=1.9.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
pytest>=8.0.0
